+++ pkgs/ffigen/lib/src/context.dart
@@ -28,7 +28,8 @@ class Context {
   final reportedCommentRanges = <((String, int), (String, int))>{};
   final libs = LibraryImports();
   late final compilerOpts =
-      config.input.compilerOptions ?? defaultCompilerOpts(logger);
+      config.input.compilerOptions ??
+      defaultCompilerOpts(logger, cpp: config.cpp != null);
   final Scope rootScope = Scope.createRoot('root');
   final Scope rootObjCScope = Scope.createRoot('objc_root');
   late final ExtraSymbols extraSymbols;